Add class visitor to annotation descriptor
As the module loaders annotation descriptors loads all java classes to
find the one annotated module class, this commit adds a class visitor to
the module annotation descriptor.
This allows users to collect other files inside the jar while the module
is loading, allowing those classes to be used as well.

To reflect this, non breaking, change the version number was increased
from 1.4.0 to 1.4.1.